LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-none-block.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/frame-ancestors-test.js"></script>
</head>
<body>
<p>A 'frame-ancestors' CSP directive with a value 'none' should block rendering. This test PASSED if you do not see any content in the iframe below. Otherwise, it FAILED.</p>
<script>
    sameOriginFrameShouldBeBlocked("'none'");
</script>
</body>
</html>
